+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 0.2.2 - Unreleased
 
 - Add changelog to the documentation and make docs deployment manually runnable.
+- Add system-aware dark mode to the web UI.
 
 ## 0.2.1 - 2026-05-29